updated single post page

+++ controllers/api/comment-routes.js
@@ -2,29 +2,33 @@ const router = require('express').Router();
 const { Comment } = require('../../models');
 
 router.get('/', (req, res) => {
-    Comment.findAll()
-      .then(dbCommentData => res.json(dbCommentData))
-      .catch(err => {
-        console.log(err);
-        res.status(500).json(err);
-      });
-  });
-
-// POST route
-router.post('/', (req, res) => {
-    Comment.create({
-        comment_text: req.body.comment_text,
-        user_id: req.body.user_id,
-        post_id: req.body.post_id
-    })
+  Comment.findAll()
     .then(dbCommentData => res.json(dbCommentData))
     .catch(err => {
-        console.log(err);
-        res.status(400).json(err);
+      console.log(err);
+      res.status(500).json(err);
     });
 });
 
+router.post('/', (req, res) => {
+  // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
+  if (req.session) {
+    Comment.create({
+      comment_text: req.body.comment_text,
+      // use the id from the session
+      user_id: req.session.user_id,
+      post_id: req.body.post_id
+    })
+      .then(dbCommentData => res.json(dbCommentData))
+      .catch(err => {
+        console.log(err);
+        res.status(400).json(err);
+      });
+  }
+});
+
 router.delete('/:id', (req, res) => {
+  if (req.session) {
     Comment.destroy({
       where: {
         id: req.params.id
@@ -41,6 +45,7 @@ router.delete('/:id', (req, res) => {
         console.log(err);
         res.status(500).json(err);
       });
-  });
+  }
+});
 
 module.exports = router;
